Select and move from true 3D fixed
Fixed the issues of a) can't select a line end unless at least one co-ordinate is at an axis limit and b) can't disambiguate 2D to 3D to allow movement of the selected line. Fixed by a) using 2D xy to select the line end (index), and b) using the *mouse* 3D coordinates instead of the line end, because the mouse 3D always does have at least one coordinate pinned to a backplane.